requirements.txt
discord.py==2.4.0
motor==3.3.2          # Compatible version for Python 3.13
pymongo==3.13.0       # Compatible version for Python 3.13
aiohttp==3.9.4        # fine
PyYAML==6.0           # fine
Pillow==10.1.0        # fine
fastapi==0.109.0      # fine
uvicorn==0.27.0       # fine
pytest==7.4.4         # fine
pytest-asyncio==0.23.3 # fine
python-dotenv==1.0.0  # fine
httpx==0.26.0         # fine
python-multipart==0.0.6 # fine
yt-dlp==2024.3.10     # fine
PyNaCl==1.5.0         # fine
